requirements.txt
numpy
vowpalwabbit
streamlit
snowflake-snowpark-python==1.3.0

